requirements.txt
coreapi==2.3.0
Django==1.11
django-appconf==1.0.1
django-extensions==1.7.8
django-filter==1.0.2
djangorestframework==3.6.2
django-rest-swagger==2.1.2
drf-nested-routers==0.11.1
ipdb==0.10.3
ipython==5.3.0
Pillow==4.1.1
pkg-resources==0.0.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
qrcode==4.0.4
raven==5.0.0
six==1.10.0
easy-thumbnails==2.4.1
Unipath==1.1
django-solo==1.1.2
django-image-cropping==1.0.3
django-debug-toolbar==1.8
celery==4.0.2
django-celery-results==1.0.1
redis==2.10.5
